+++ MatrixSDK/Data/MXRoomSummary.h
@@ -43,6 +43,8 @@
  */
 FOUNDATION_EXPORT NSString *const kMXRoomSummaryDidChangeNotification;
 
+/// Number of events retrieved when doing pagination from the homeserver.
+FOUNDATION_EXPORT NSUInteger const MXRoomSummaryPaginationChunkSize;
 
 /**
  `MXRoomSummary` exposes and caches data for a room.
@@ -247,18 +249,39 @@ FOUNDATION_EXPORT NSString *const kMXRoomSummaryDidChangeNotification;
 @property (nonatomic) MXEvent *lastMessageEvent;
 
 /**
- Reset the last message.
+ Intenal SDK method to load and decrypt the MXEvent of the last message.
  
- The operation is asynchronous as it may require pagination from the homeserver.
+ @param onComplete the callback called once operation is done.
+ */
+-(void)loadLastEvent:(void (^)(void))onComplete;
+
+/**
+ Reset the last message from data in the store.
  
- @param complete A block object called when the operation completes.
+ @param onComplete A block object called when the operation completes.
  @param failure A block object called when the operation fails.
  @param commit  Tell whether the updated room summary must be committed to the store. Use NO when a more
  global [MXStore commit] will happen. This optimises IO.
 
  @return a MXHTTPOperation instance.
  */
-- (MXHTTPOperation*)resetLastMessage:(void (^)(void))complete failure:(void (^)(NSError *))failure commit:(BOOL)commit;
+- (MXHTTPOperation*)resetLastMessage:(void (^)(void))onComplete failure:(void (^)(NSError *))failure commit:(BOOL)commit;
+
+/**
+ Reset the last message by paginating messages from the homeserver if needed.
+ 
+ @param maxServerPaginationCount The max number of messages to retrieve from the server.
+ @param onComplete A block object called when the operation completes.
+ @param failure A block object called when the operation fails.
+ @param commit  Tell whether the updated room summary must be committed to the store. Use NO when a more
+ global [MXStore commit] will happen. This optimises IO.
+ 
+ @return a MXHTTPOperation instance.
+ */
+- (MXHTTPOperation *)resetLastMessageWithMaxServerPaginationCount:(NSUInteger)maxServerPaginationCount
+                                                       onComplete:(void (^)(void))onComplete
+                                                          failure:(void (^)(NSError *))failure
+                                                           commit:(BOOL)commit;
 
 
 #pragma mark - Data related to business logic
@@ -350,8 +373,9 @@ FOUNDATION_EXPORT NSString *const kMXRoomSummaryDidChangeNotification;
  Note: state events have been previously sent to `handleStateEvents`.
 
  @param roomSync information to sync the room with the home server data.
+ @param onComplete the block called when the operation completes.
  */
-- (void)handleJoinedRoomSync:(MXRoomSync*)roomSync;
+- (void)handleJoinedRoomSync:(MXRoomSync*)roomSync onComplete:(void (^)(void))onComplete;
 
 /**
  Update the invited room state according to the provided data.