+++ MatrixSDK/VoIP/MXCallManager.m
@@ -117,6 +117,8 @@ NSTimeInterval const kMXCallDirectRoomJoinTimeout = 30;
                                                                 kMXEventTypeStringCallNegotiate,
                                                                 kMXEventTypeStringCallReplaces,
                                                                 kMXEventTypeStringCallRejectReplacement,
+                                                                kMXEventTypeStringCallAssertedIdentity,
+                                                                kMXEventTypeStringCallAssertedIdentityUnstable,
                                                                 kMXEventTypeStringRoomMember
                                                                 ]
                                                       onEvent:^(MXEvent *event, MXTimelineDirection direction, id customObject) {
@@ -394,6 +396,10 @@ NSTimeInterval const kMXCallDirectRoomJoinTimeout = 30;
         case MXEventTypeCallRejectReplacement:
             [self handleCallRejectReplacement:event];
             break;
+        case MXEventTypeCallAssertedIdentity:
+        case MXEventTypeCallAssertedIdentityUnstable:
+            [self handleCallAssertedIdentity:event];
+            break;
         case MXEventTypeRoomMember:
             [self handleRoomMember:event];
             break;
@@ -632,6 +638,55 @@ NSTimeInterval const kMXCallDirectRoomJoinTimeout = 30;
     }
 }
 
+- (void)handleCallAssertedIdentity:(MXEvent *)event
+{
+    //  check handling allowed
+    if (![MXSDKOptions sharedInstance].handleCallAssertedIdentityEvents)
+    {
+        return;
+    }
+    
+    MXCallAssertedIdentityEventContent *content = [MXCallAssertedIdentityEventContent modelFromJSON:event.content];
+    
+    // Forward the event to the MXCall object
+    MXCall *call = [self callWithCallId:content.callId];
+    if (call)
+    {
+        if (content.assertedIdentity.userId)
+        {
+            //  do a native lookup first
+            
+            MXWeakify(self);
+            
+            [self getNativeUserFrom:content.assertedIdentity.userId success:^(MXThirdPartyUserInstance * _Nonnull user) {
+                MXStrongifyAndReturnIfNil(self);
+                
+                MXAssertedIdentityModel *assertedIdentity = content.assertedIdentity;
+                
+                //  fetch the native user
+                MXUser *mxUser = [self.mxSession userWithUserId:user.userId];
+                
+                if (mxUser)
+                {
+                    assertedIdentity = [[MXAssertedIdentityModel alloc] initWithUser:mxUser];
+                }
+                else
+                {
+                    assertedIdentity.userId = user.userId;
+                }
+                
+                //  use the updated asserted identity
+                call.assertedIdentity = assertedIdentity;
+            } failure:nil];
+        }
+        else
+        {
+            //  no need to a native lookup, directly pass the identity
+            call.assertedIdentity = content.assertedIdentity;
+        }
+    }
+}
+
 - (void)handleRoomMember:(MXEvent *)event
 {
     MXRoomMemberEventContent *content = [MXRoomMemberEventContent modelFromJSON:event.content];
@@ -750,7 +805,7 @@ NSTimeInterval const kMXCallDirectRoomJoinTimeout = 30;
                   to:(MXUserModel *)target
       withTransferee:(MXUserModel *)transferee
         consultFirst:(BOOL)consultFirst
-             success:(void (^)(NSString * _Nonnull newCallId))success
+             success:(void (^)(NSString * _Nullable newCallId))success
              failure:(void (^)(NSError * _Nullable error))failure
 {
     if (callWithTransferee.isConferenceCall)
@@ -790,19 +845,56 @@ NSTimeInterval const kMXCallDirectRoomJoinTimeout = 30;
                     return;
                 }
                 
-                //  generate a new call id
-                NSString *newCallId = [[NSUUID UUID] UUIDString];
-                
-                //  first, send replaces event to the call with the transferee, send info about target
-                //  to make transferee start a new call
-                [callWithTransferee transferToRoom:transferRoom.roomId
-                                              user:target
-                                        createCall:newCallId
-                                         awaitCall:nil
-                                           success:^(NSString * _Nonnull eventId) {
+                if (consultFirst)
+                {
+                    //  consult with the target
+                    if (callWithTarget.isOnHold)
+                    {
+                        [callWithTarget hold:NO];
+                    }
                     
-                    //  define block to be called after replaces events are sent
-                    void(^afterReplacesEventsBlock)(void) = ^{
+                    if (success)
+                    {
+                        success(nil);
+                    }
+                }
+                else
+                {
+                    //  generate a new call id
+                    NSString *newCallId = [[NSUUID UUID] UUIDString];
+                    
+                    dispatch_group_t dispatchGroupReplaces = dispatch_group_create();
+                    
+                    if (callWithTarget)
+                    {
+                        [callWithTarget hangup];
+                        
+                        //  send replaces event to target
+                        dispatch_group_enter(dispatchGroupReplaces);
+                        [callWithTarget transferToRoom:transferRoom.roomId
+                                                  user:transferee
+                                            createCall:nil
+                                             awaitCall:newCallId
+                                               success:^(NSString * _Nonnull eventId) {
+                            dispatch_group_leave(dispatchGroupReplaces);
+                        } failure:failure];
+                    }
+                    
+                    dispatch_group_enter(dispatchGroupReplaces);
+                    if (callWithTransferee.isOnHold)
+                    {
+                        [callWithTransferee hold:NO];
+                    }
+                    //  send replaces event to transferee
+                    [callWithTransferee transferToRoom:transferRoom.roomId
+                                                  user:target
+                                            createCall:newCallId
+                                             awaitCall:nil
+                                               success:^(NSString * _Nonnull eventId) {
+                        dispatch_group_leave(dispatchGroupReplaces);
+                    } failure:failure];
+                    
+                    dispatch_group_notify(dispatchGroupReplaces, dispatch_get_main_queue(), ^{
                         if (isNewRoom)
                         {
                             //  if was a newly created room, send invites after replaces events
@@ -810,58 +902,25 @@ NSTimeInterval const kMXCallDirectRoomJoinTimeout = 30;
                             [transferRoom inviteUser:transferee.userId success:nil failure:failure];
                         }
                         
-                        //  do not end the calls, wait for other parties to end it
-                        
-                        //  TODO: Observe the call with the newCallId
-                        
                         if (success)
                         {
                             success(newCallId);
                         }
-                    };
-                    
-                    if (callWithTarget)
-                    {
-                        //  then, send replaces event to the call with the target, send info about transferee
-                        //  to make target await a call
-                        [callWithTarget transferToRoom:transferRoom.roomId
-                                                  user:transferee
-                                            createCall:nil
-                                             awaitCall:newCallId
-                                               success:^(NSString * _Nonnull eventId) {
-                            
-                            if (consultFirst)
-                            {
-                                //  hold the callWithTransferee
-                                [callWithTransferee hold:YES];
-                                
-                                //  consult with the target
-                                [callWithTarget hold:NO];
-                                
-                                //  TODO: Make a state change on callWithTransferee here, like `onHoldForConsulting` and provide target details, to complete the transfer after
-                            }
-                            else
-                            {
-                                //  no need to consult
-                                
-                                afterReplacesEventsBlock();
-                            }
-                            
-                        } failure:failure];
-                    }
-                    else
-                    {
-                        //  being here means no need to consult, otherwise would fail before
-                        
-                        afterReplacesEventsBlock();
-                    }
-                    
-                } failure:failure];
+                    });
+                }
+                
             }];
         };
         
         if (call)
         {
+            if (consultFirst)
+            {
+                call.callWithTransferee = callWithTransferee;
+                call.transferee = transferee;
+                call.transferTarget = target;
+                call.consulting = YES;
+            }
             continueBlock(call);
         }
         else
@@ -886,15 +945,22 @@ NSTimeInterval const kMXCallDirectRoomJoinTimeout = 30;
                         return;
                     }
                     
-                    //  hold the transferee call before starting the new call
-                    [callWithTransferee hold:YES];
-                    
-                    //  place a new call to the target
-                    [self placeCallInRoom:room.roomId withVideo:callWithTransferee.isVideoCall success:^(MXCall * _Nonnull call) {
+                    //  place a new audio call to the target to consult the transfer
+                    [self placeCallInRoom:room.roomId withVideo:NO success:^(MXCall * _Nonnull call) {
+                        
+                        //  mark the call with target & transferee as consulting
+                        call.callWithTransferee = callWithTransferee;
+                        call.transferee = transferee;
+                        call.transferTarget = target;
+                        call.consulting = YES;
                         
                         continueBlock(call);
                     } failure:^(NSError * _Nullable error) {
                         NSLog(@"[MXCallManager] transferCall: couldn't call the target: %@", error);
+                        if (failure)
+                        {
+                            failure(error);
+                        }
                     }];
                     
                 }];