+++ CHANGES.rst
@@ -5,14 +5,46 @@ Changes to be released in next version
  * Spaces: Support Space room type (vector-im/element-ios/issues/4069).
 
 🙌 Improvements
- * MXHTTPOperation: Expose the HTTP response (vector-im/element-ios/issues/4206).
- * MXRoomPowerLevels: Handle undefined values and add init with default spec values.
- * MXRoomCreationParameters: Add roomType and powerLevelContentOverride properties. Add initial state events update method.
- * MXResponse: Add convenient uncurry method to convert a Swift method into Objective-C.
- * Add MXRoomInitialStateEventBuilder that enables to build initial state events.
+ * MXSession: Cache initial sync response until it is fully handled (vector-im/element-ios/issues/4317).
+ * MXStore: New commit method accepting a completion block.
+ * MXCrypto: Decrypt events asynchronously and no more on the main thread )(vector-im/element-ios/issues/4306).
+ * MXSession: Add the decryptEvents method to decypt a bunch of events asynchronously.
+ * MXSession: Make the eventWithEventId method decrypt the event if needed.
+ * MXEventTimeline: Add NSCopying implementation so that another pagination can be done on the same set of data.
+ * MXCrypto: eventDeviceInfo: Do not synchronise anymore the operation with the decryption queue.
+ * MXRoomSummary: Improve reset resetLastMessage to avoid pagination loop and to limit number of decryptions.
+ * MXSession: Limit the number of decryptions when processing an initial sync (vector-im/element-ios/issues/4307).
 
 🐛 Bugfix
+ * MXRoomSummary: Fix decryption of the last message when it is edited (vector-im/element-ios/issues/4322).
+ * MXCall: Check remote partyId for select_answer events (vector-im/element-ios/issues/4337).
+
+⚠️ API Changes
+ * MXRoom: MXRoom.outgoingMessages does not decrypt messages anymore. Use MXSession.decryptEvents to get decrypted events.
+ * MXSession: [MXSession decryptEvent:inTimeline:] is deprecated, use [MXSession decryptEvents:inTimeline:onComplete:] instead.
+ * MXCrypto: [MXCrypto decryptEvent:inTimeline:] is deprecated, use [MXCrypto decryptEvents:inTimeline:onComplete:] instead.
+ * MXCrypto: [MXCrypto hasKeysToDecryptEvent:] is now asynchronous.
+
+🗣 Translations
  * 
+    
+🧱 Build
+ * 
+
+Others
+ * 
+
+Changes in 0.18.12 (2021-05-12)
+=================================================
+
+✨ Features
+ * 
+
+🙌 Improvements
+ * MXPushGatewayRestClient: Add timeout param to the HTTP method.
+
+🐛 Bugfix
+ * MXRoomCreateContent: Fix room type JSON key.
 
 ⚠️ API Changes
  * 
@@ -25,6 +57,76 @@ Changes to be released in next version
 
 Others
  * 
+
+Improvements:
+
+
+Changes in 0.18.11 (2021-05-07)
+=================================================
+
+✨ Features
+ * 
+
+🙌 Improvements
+ * MXCallKitAdapter: Update incoming calls if answered from application UI.
+ * MXFileStore: Logs all files when a data corruption is detected (to track vector-im/element-ios/issues/4921).
+ * MXCallManager: Fix call transfers flow for all types of transfers.
+ * VoIP: Implement asserted identity for calls: MSC3086 (matrix-org/matrix-doc/pull/3086).
+
+🐛 Bugfix
+ * MXTools: Fix bad linkification of matrix alias and URL (vector-im/element-ios/issues/4258).
+ * MXRoomSummary: Fix roomType property deserialization issue.
+ * MXCall: Disable call transferee capability & fix call transfer feature check.
+
+⚠️ API Changes
+ * Spaces and room type: Remove all MSC1772 JSON key prefixes and use stable ones.
+
+🗣 Translations
+ * 
+    
+🧱 Build
+ * Tests: Use UnitTests suffix for unit tests classes.
+ * Tests: Cut some existing tests to separate unit tests and integration tests.
+ * Tests: Create 4 test plans for the macOS target: AllTests, AllTestsWithSanitizers, UnitTests and UnitTestsWithSanitizers.
+ * GH Actions: Run unit tests on every PR and develop branch update.
+ * GH Actions: Run integration tests nightly on develop using last Synapse release.
+
+Others
+ * 
+
+Improvements:
+
+
+Changes in 0.18.10 (2021-04-22)
+=================================================
+
+✨ Features
+ * 
+
+🙌 Improvements
+ * MXHTTPOperation: Expose the HTTP response (vector-im/element-ios/issues/4206).
+ * MXRoomPowerLevels: Handle undefined values and add init with default spec values.
+ * MXRoomCreationParameters: Add roomType and powerLevelContentOverride properties. Add initial state events update method.
+ * MXResponse: Add convenient uncurry method to convert a Swift method into Objective-C.
+ * Add MXRoomInitialStateEventBuilder that enables to build initial state events.
+
+🐛 Bugfix
+ * MXCrypto: Disable optimisation on room members list to make sure we share keys to all (vector-im/element-ios/issues/3807).
+
+⚠️ API Changes
+ * 
+
+🗣 Translations
+ * 
+    
+🧱 Build
+ * 
+
+Others
+ * 
+
+Improvements:
+
 
 Changes in 0.18.9 (2021-04-16)
 =================================================